rename buildGeolocationParameter into buildLocationParameter
