+++ src/libs/algolia/fetchAlgolia/buildAlgoliaParameters/buildLocationParameter.ts
